mdpdf

+++ hack/mdpdf/convert.sh
@@ -1,9 +1,17 @@
 #!/bin/bash
 
-mdpdf -o ../../docs/nkia-api/api.pdf ../../docs/nkia-api/index.md 
+if [ ! -d "VENV" ]; then
+  echo "VENV does not exist."
+  exit -1
+fi
+
+source VENV/bin/activate
 
 
-mdpdf -o ../../docs/nkia-server/server.pdf ../../docs/nkia-server/index.md 
+mdpdf -o ../../doc/nkia-api/api.pdf ../../doc/nkia-api/index.md 
+
+
+mdpdf -o ../../doc/nkia-server/server.pdf ../../doc/nkia-server/index.md 
 
 
 rm mdpdf.log
